feat(binaries): allow many binaries to be registered for the same name (#141)
This feature bring the ability to have the same command in different version
inside the same project.

This add the `condition` parameter on binaries so you can restrict the
availability of a command inside a given project directory or configuration
option.

When a binary shim is invoked, it runs `ddb run <name>` under the hood.
All Binaries matching this name with a condition defined will have their
condition evaluated, and first condition returning True will be used to
generate the effective command output. If no condition match, it will
use the first of remaining Binary matching this name (without condition).
If no binary matching this name is finally found, it will invoke the command
wrapped into $(ddb deactivate)/$(ddb activate) instruction, so the command
run on the host system.

Close #141